backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
httpx==0.28.1
eval_type_backport==0.2.2
python-multipart==0.0.20
Pillow==11.3.0
openai==1.102.0
python-dotenv==1.0.1
